// web-panel/src/api/models/agentSpecs.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Monitoring Server API
 * OpenAPI spec version: 1.0
 */
import type { CpuSpecs } from './cpuSpecs';
import type { DiskSpecs } from './diskSpecs';
import type { HostSpecs } from './hostSpecs';
import type { MemorySpecs } from './memorySpecs';

export interface AgentSpecs {
  agentId?: string;
  cpuSpecs?: CpuSpecs;
  diskSpecs?: DiskSpecs[];
  hostSpecs?: HostSpecs;
  memorySpecs?: MemorySpecs;
  updatedAt?: string;
}